requirements.txt
openai>=1.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0
requests>=2.25.0
tenacity>=8.0.0
PyYAML>=6.0
Pillow>=10.0.0
pillow-heif>=0.10.0
